Fix exception when trying to connect to unity3d.com over HTTPS
Unfortunately, it’s no longer possible to use the Python bundled with macOS. Unity changed their server configuration so that TLS1.2 is required to connect. macOS ships with OpenSSL version 0.9.8 or lower, which only supports TLS1.0. To connect to unity3d.com, a new version of python that is built using OpenSSL 1.0.1.g or newer is required. The script now checks the OpenSSL version and prompts the user to update python if it’s too old.