Fix installation when running as root
Issue is that tar changes many default flags when running as root and tries to restore much more from the archive. The main editor package contains an entry for '.', which maps to '/Applications'. Tar then tries to update the owner of '/Application', which is prevented by SIP.
Add a bunch of flags which are on by default when running as non-root but are off when running as root.